fix: リファクタとSenderの生存期間をエイっとしました
