+++ ytdlpSpout/constants.py
@@ -1,0 +1,16 @@
+"""ytdlpSpout 共通定数定義"""
+
+# デフォルト設定
+DEFAULT_VIDEO_URL = "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
+DEFAULT_SENDER_NAME = "ytdlpSpoutSender"
+DEFAULT_WIDTH = 1920
+DEFAULT_HEIGHT = 1080
+DEFAULT_FPS = 30
+
+# 連続失敗の最大回数
+MAX_CONSECUTIVE_FAILURES = 3
+
+# FPS制限
+MIN_FPS = 15
+MAX_FPS = 120
+
